requirements.txt
asgiref==3.8.1
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
